+++ base/src/main/java/module-info.java
@@ -6,6 +6,7 @@ module atlantafx.base {
     requires static org.jetbrains.annotations;
 
     exports atlantafx.base.controls;
+    exports atlantafx.base.layout;
     exports atlantafx.base.theme;
     exports atlantafx.base.util;