+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 
 ### Features
 
+- (Base) New `DeckPane` component with swipe and slide transition support.
 - (CSS) 🚀 New MacOS-like Cupertino theme in light and dark variants.
 - (CSS) 🚀 New [Dracula](https://ui.draculatheme.com/) theme.
-- (CSS) 🚀 New `TabPane` style. There are three styles supported: default, floating and classic (new one).
+- (CSS) New `TabPane` style. There are three styles supported: default, floating and classic (new one).
+- (Sampler) 🚀 SceneBuilder integration. AtlantaFX themes can be installed (or uninstalled) to SceneBuilder directly from the Sampler app. 
 
 ### Improvements